dom/base/test/test_bug704063.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug  **/
  SimpleTest.waitForExplicitFinish();
  var firstRan = false;
  var secondRan = false;
  function second(time) {
    is(firstRan, true, "We were called second");
    secondRan = true;
    ok(Math.abs(time - performance.now()) < 3600000,
       "An hour really shouldn't have passed here");
    ok(Math.abs(time - Date.now()) > 3600000,
       "More than an hour should have passed since 1970");
  }
  function first(time) {
    is(secondRan, false, "second() was called first");
    firstRan = true;
    ok(Math.abs(time - performance.now()) < 3600000,
       "An hour really shouldn't have passed here either");
    ok(Math.abs(time - Date.now()) > 3600000,
       "More than an hour should have passed since 1970 here either");
  }
  function third() {
    ok(firstRan, "We should be after the first call");
    ok(secondRan, "We should be after the second call");
    SimpleTest.finish();
  }

  window.onload = function() {
    window.requestAnimationFrame(first);
    window.requestAnimationFrame(second);
    window.requestAnimationFrame(third);
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
